moved Box component to seperate file via components directory
