+++ src/App.js
@@ -1,17 +1,7 @@
-import React, { useRef } from "react";
+import React from "react";
 import "./App.scss";
-import { Canvas, useFrame } from "@react-three/fiber";
-
-const Box = () => {
-  const mesh = useRef(null);
-  useFrame(() => (mesh.current.rotation.x = mesh.current.rotation.y += 0.01));
-  return (
-    <mesh ref={mesh}>
-      <boxBufferGeometry attach="geometry" args={[1, 1, 1]} />
-      <meshStandardMaterial attach="material" />
-    </mesh>
-  );
-};
+import { Canvas } from "@react-three/fiber";
+import Box from "./components/Box";
 
 function App() {
   return (
